requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# LINE Bot
# No SDK - using direct REST API (line_service.py)

# OCR
paddleocr>=3.4.0
# PaddlePaddle installed separately per platform

# AI/LLM
# google-generativeai not needed - using REST API directly

# Excel
openpyxl>=3.1.0

# PDF (for handler certificate)
reportlab>=4.0

# Testing
pytest>=9.0.0
